requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Deprecated==1.2.10
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
Pillow==8.1.0
PyGithub==1.54.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
wrapt==1.12.1
